@@ -13,7 +13,7 @@ from ... import constants as rpc
 from .base import AgentSchedulingComponent
 
 
-# -----------------------------------------------------------------------------=
+# ------------------------------------------------------------------------------
 #
 class Spark(AgentSchedulingComponent):
 
